requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.6.15
click==8.2.1
exceptiongroup==1.3.0
fastapi==0.115.14
graphql-core==3.2.6
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
packaging==25.0
pluggy==1.6.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pytest==8.4.1
pytest-asyncio==1.0.0
python-dateutil==2.9.0.post0
python-multipart==0.0.20
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
strawberry-graphql==0.275.5
tomli==2.2.1
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
